fix: linter fix, package restructure with docs bump
