+++ docs/resources/graph_beta_device_management_windows_feature_update_profile.md
@@ -174,7 +174,7 @@ resource "microsoft365_graph_beta_device_management_windows_feature_update_profi
 ### Required
 
 - `display_name` (String) The display name of the profile.
-- `feature_update_version` (String) The feature update version that will be deployed to the devices targeted by this profile. Valid values are: "Windows 11, version 24H2", "Windows 11, version 23H2", "Windows 11, version 22H2", "Windows 10, version 22H2". By selecting this Feature update to deploy you are agreeing that when applying this operating system to a device either (1) the applicable Windows license was purchased though volume licensing, or (2) that you are authorized to bind your organization and are accepting on its behalf the relevant Microsoft Software License Terms to be found here https://go.microsoft.com/fwlink/?linkid=2171206.
+- `feature_update_version` (String) The feature update version that will be deployed to the devices targeted by this profile. Valid values are: "Windows 11, version 25H2", "Windows 11, version 24H2", "Windows 11, version 23H2", "Windows 11, version 22H2", "Windows 10, version 22H2". By selecting this Feature update to deploy you are agreeing that when applying this operating system to a device either (1) the applicable Windows license was purchased though volume licensing, or (2) that you are authorized to bind your organization and are accepting on its behalf the relevant Microsoft Software License Terms to be found here https://go.microsoft.com/fwlink/?linkid=2171206.
 
 ### Optional
