+++ docs/resources/graph_identity_and_access_conditional_access_terms_of_use.md
@@ -1,0 +1,84 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "microsoft365_graph_identity_and_access_conditional_access_terms_of_use Resource - terraform-provider-microsoft365"
+subcategory: ""
+description: |-
+  Manages Microsoft 365 Terms of Use Agreements using the /agreements endpoint. Terms of use agreements allow organizations to present information that users must accept before accessing data or applications. These agreements can be used to ensure compliance with legal or regulatory requirements.
+---
+
+# microsoft365_graph_identity_and_access_conditional_access_terms_of_use (Resource)
+
+Manages Microsoft 365 Terms of Use Agreements using the `/agreements` endpoint. Terms of use agreements allow organizations to present information that users must accept before accessing data or applications. These agreements can be used to ensure compliance with legal or regulatory requirements.
+
+
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `display_name` (String) Display name of the agreement. The display name is used for internal tracking of the agreement but isn't shown to end users who view the agreement.
+- `file` (Attributes) Default PDF linked to this agreement. This is required when creating a new agreement. (see [below for nested schema](#nestedatt--file))
+- `is_per_device_acceptance_required` (Boolean) This setting enables you to require end users to accept this agreement on every device that they're accessing it from. The end user is required to register their device in Microsoft Entra ID, if they haven't already done so.
+- `is_viewing_before_acceptance_required` (Boolean) Indicates whether the user has to expand the agreement before accepting.
+
+### Optional
+
+- `terms_expiration` (Attributes) Expiration schedule and frequency of agreement for all users. (see [below for nested schema](#nestedatt--terms_expiration))
+- `timeouts` (Attributes) (see [below for nested schema](#nestedatt--timeouts))
+- `user_reaccept_required_frequency` (String) The duration after which the user must reaccept the terms of use. Accepted values: `P365D` for annually, `P180D` for bi-annually, `P90D` for quarterly, `P30D` for monthly.
+
+### Read-Only
+
+- `id` (String) The unique identifier for the agreement. This is automatically generated when the agreement is created.
+
+<a id="nestedatt--file"></a>
+### Nested Schema for `file`
+
+Required:
+
+- `localizations` (Attributes Set) The localized version of the terms of use agreement files attached to the agreement. (see [below for nested schema](#nestedatt--file--localizations))
+
+<a id="nestedatt--file--localizations"></a>
+### Nested Schema for `file.localizations`
+
+Required:
+
+- `display_name` (String) Localized display name of the policy file of an agreement. The localized display name is shown to end users who view the agreement.
+- `file_data` (Attributes) Data that represents the terms of use PDF document. (see [below for nested schema](#nestedatt--file--localizations--file_data))
+- `file_name` (String) Name of the agreement file (for example, TOU.pdf).
+- `language` (String) The language of the agreement file in the format 'languagecode2-country/regioncode2'. 'languagecode2' is a lowercase two-letter code derived from ISO 639-1, while 'country/regioncode2' is derived from ISO 3166 and usually consists of two uppercase letters, or a BCP-47 language tag. For example, U.S. English is `en-US`.
+
+Optional:
+
+- `is_default` (Boolean) If none of the languages matches the client preference, indicates whether this is the default agreement file. If none of the files are marked as default, the first one is treated as the default. Read-only.
+- `is_major_version` (Boolean) Indicates whether the agreement file is a major version update. Major version updates invalidate the agreement's acceptances on the corresponding language.
+
+<a id="nestedatt--file--localizations--file_data"></a>
+### Nested Schema for `file.localizations.file_data`
+
+Required:
+
+- `data` (String) Data that represents the terms of use PDF document as raw bytes.
+
+
+
+
+<a id="nestedatt--terms_expiration"></a>
+### Nested Schema for `terms_expiration`
+
+Optional:
+
+- `frequency` (String) Represents the frequency at which the terms will expire, after its first expiration as set in startDateTime. Accepted values: `P365D` for annually, `P180D` for bi-annually, `P90D` for quarterly, `P30D` for monthly.
+- `start_date_time` (String) The DateTime when the agreement is set to expire for all users. The Timestamp type represents date and time information using ISO 8601 format and is always in UTC time.
+
+
+<a id="nestedatt--timeouts"></a>
+### Nested Schema for `timeouts`
+
+Optional:
+
+- `create` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours).
+- `delete` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours). Setting a timeout for a Delete operation is only applicable if changes are saved into state before the destroy operation occurs.
+- `read` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours). Read operations occur during any refresh or planning operation when refresh is enabled.
+- `update` (String) A string that can be [parsed as a duration](https://pkg.go.dev/time#ParseDuration) consisting of numbers and unit suffixes, such as "30s" or "2h45m". Valid time units are "s" (seconds), "m" (minutes), "h" (hours).
