+++ templates/ephemeral/graph_beta_multitenant_management_audit_event.md.tmpl
@@ -1,0 +1,37 @@
+---
+page_title: "{{.Name}} {{.Type}} - {{.ProviderName}}"
+subcategory: "Multitenant Management"
+
+description: |-
+{{ if .Description }}{{ .Description | plainmarkdown | trimspace | prefixlines "  " }}{{ else }}  Retrieves audit events from Microsoft 365 managed tenants as an ephemeral resource.{{ end }}
+---
+
+# {{.Name}} ({{.Type}})
+
+{{ if .Description }}{{ .Description | trimspace }}{{ else }}Retrieves audit events from Microsoft 365 managed tenants as an ephemeral resource. This does not persist in state and fetches fresh data on each execution.{{ end }}
+
+## Microsoft Documentation
+
+- [auditEvent resource type](https://learn.microsoft.com/en-us/graph/api/resources/managedtenants-auditevent?view=graph-rest-beta)
+
+## API Permissions
+
+The following API permissions are required in order to use this data source.
+
+### Microsoft Graph
+
+- **Application**: `ManagedTenants.Read.All`, `ManagedTenants.ReadWrite.All`.
+
+## Version History
+
+| Version | Status | Notes |
+|---------|--------|-------|
+| v0.31.0-alpha | Experimental | Initial release |
+
+## Important Notes
+This method supports some of the OData query parameters to help customize the response. For general information, see OData query parameters.
+## Example Usage
+
+{{ tffile "examples/ephemerals/microsoft365_graph_beta_multitenant_management_audit_event/ephemeral.tf" }}
+
+{{ if .SchemaMarkdown }}{{ .SchemaMarkdown | trimspace }}{{ else }}Schema documentation not available.{{ end }}
